Update README.md
This is text based italian card game named "Briscola".
This is version for two players only. (Computer vs player)

Rules:

After the deck is shuffled, each player is dealt three cards. The next card is placed face up on the playing surface, and the remaining deck is placed face down, sometimes covering half of the up-turned card. This card is the Briscola, and represents the trump suit for the game. The deal, and game play itself, proceeds counter-clockwise.
The winner of that hand is determined as follows:

if any briscola (trump) has been played, the player who played the highest valued trump wins
if no briscole (trumps) have been played, the player who played the highest card of the lead suit wins
Unlike other trump card games, players are not required to follow suit, that is, to play the same suit as the lead player.

Once the winner of a trick is determined, that player collects the played cards, and places them face down in a pile. Each player maintains his/her own pile. Then, each player draws a card from the remaining deck, starting with the player who won the trick, proceeding counter-clockwise. Note that the last card collected in the game should be the up-turned Briscola. The player who won the trick leads the next hand. During game play and only before the next to the last hand is played, a player who draws the card with the seven of trump can take the "briscola". This may be done only if the player has won a hand. Before the last hand, people in the same team can look at each other's cards.

After all cards have been played, players calculate the total point value of cards in their own piles. For multi-player games, partners combine their points.